+++ src/argus/htmx/auth/urls.py
@@ -1,0 +1,11 @@
+from django.contrib.auth import views as django_auth_views
+from django.urls import path
+
+from . import views as auth_views
+
+
+urlpatterns = [
+    path("accounts/login/", auth_views.LoginView.as_view(), name="login"),
+    path("accounts/logout/", django_auth_views.LogoutView.as_view(), name="logout"),
+    # path("accounts/", include("django.contrib.auth.urls")),
+]
