+++ src/argus/htmx/auth/views.py
@@ -5,13 +5,11 @@ from django.urls import reverse
 from argus.auth.utils import (
     has_model_backend,
     has_remote_user_backend,
-    get_psa_authentication_backends,
     get_authentication_backend_classes,
 )
 
 
 REMOTE_USER_METHOD_NAME = getattr(settings, "ARGUS_REMOTE_USER_METHOD_NAME", "REMOTE_USER")
-OIDC_METHOD_NAME = getattr(settings, "ARGUS_OIDC_METHOD_NAME", "OIDC")
 
 
 def get_htmx_authentication_backend_name_and_type():
@@ -21,7 +19,7 @@ def get_htmx_authentication_backend_name_and_type():
     data = {}
     if has_model_backend(backends):
         data["local"] = {
-            "url": reverse("htmx:login"),
+            "url": reverse("login"),
             "display_name": "Log In",
         }
 
@@ -31,17 +29,6 @@ def get_htmx_authentication_backend_name_and_type():
             "display_name": REMOTE_USER_METHOD_NAME,
         }
         data.setdefault("external", []).append(remote_user_data)
-
-    for backend in get_psa_authentication_backends(backends):
-        display_name = backend.name
-        if backend.name == "oidc":
-            display_name = OIDC_METHOD_NAME
-        psa_backend_data = {
-            "url": reverse("social:begin", kwargs={"backend": backend.name}),
-            "display_name": display_name,
-        }
-        data.setdefault("external", []).append(psa_backend_data)
-
     return data
