+++ src/argus/htmx/htmx_urls.py
@@ -1,10 +1,12 @@
-from django.urls import include, path
-
-from argus.site.utils import get_urlpatterns
-
 from argus.htmx.appconfig import APP_SETTINGS
+from argus.site.utils import get_urlpatterns
+from argus.util.app_utils import is_using_psa
 
 urlpatterns = get_urlpatterns(APP_SETTINGS)
-urlpatterns += [
-    path("oidc/", include("social_django.urls", namespace="social")),
-]
+
+if is_using_psa():
+    from argus.auth.psa.urls import urlpatterns as account_urlpatterns
+else:
+    from argus.htmx.auth.urls import urlpatterns as account_urlpatterns
+
+urlpatterns = account_urlpatterns + urlpatterns
