+++ src/argus/htmx/settings.py
@@ -12,3 +12,7 @@ LOGIN_URL = "/accounts/login/"
 LOGOUT_URL = "/accounts/logout/"
 LOGIN_REDIRECT_URL = "/incidents/"
 LOGOUT_REDIRECT_URL = "/incidents/"
+
+if USE_PYTHON_SOCIAL_AUTH:
+    INSTALLED_APPS += ["social_django"]
+    from argus.auth.psa.settings import *  # noqa: F403
