+++ src/argus/htmx/urls.py
@@ -1,7 +1,5 @@
-from django.contrib.auth import views as django_auth_views
 from django.urls import path, include
 
-from .auth import views as auth_views
 from .incident.urls import urlpatterns as incident_urls
 from .timeslot.urls import urlpatterns as timeslot_urls
 from .notificationprofile.urls import urlpatterns as notificationprofile_urls
@@ -9,11 +7,9 @@ from .destination.urls import urlpatterns as destination_urls
 from .user.urls import urlpatterns as user_urls
 from .views import IncidentListRedirectView, StyleGuideView
 
+
 app_name = "htmx"
 urlpatterns = [
-    path("accounts/login/", auth_views.LoginView.as_view(), name="login"),
-    path("accounts/logout/", django_auth_views.LogoutView.as_view(), name="logout"),
-    # path("accounts/", include("django.contrib.auth.urls")),
     path("styleguide/", StyleGuideView.as_view(), name="styleguide"),
     path("incidents/", include(incident_urls)),
     path("timeslots/", include(timeslot_urls)),
